+++ .eslintrc.js
@@ -13,7 +13,7 @@ module.exports = {
         },
     },
     parser: "@typescript-eslint/parser",
-    plugins: ["jsx-a11y"],
+    plugins: ["prettier", "jsx-a11y"],
     extends: [
         "plugin:react/recommended",
         "plugin:@typescript-eslint/recommended",
@@ -29,5 +29,7 @@ module.exports = {
         "@typescript-eslint/explicit-function-return-type": 0, // Explicit function return type becomes very annoying for React Function Components
         "@typescript-eslint/no-use-before-define": ["error", { functions: false }],
         "@typescript-eslint/ban-ts-ignore": 0, // We use ts-ignore for modules that don't have type definition files
+        "react/prop-types": 0,
+        "prettier/prettier": "error",
     },
 };