+++ packages/core/src/index.ts
@@ -1,1 +1,8 @@
-export { initTabListener } from "./utils";
+export { initTabListener, getValuePair } from "./utils";
+export { ScreenReaderOnly } from "./components";
+
+export type LabelVariant = "small" | "medium" | "large";
+export type ValuePair = {
+    value: string;
+    label: string;
+};