+++ packages/field-group-react/src/FieldGroup.tsx
@@ -1,5 +1,6 @@
 import React, { ReactNode } from "react";
-import { SupportLabel } from "@fremtind/jkl-typography-react";
+import { Label, SupportLabel } from "@fremtind/jkl-typography-react";
+import { LabelVariant } from "@fremtind/jkl-core";
 
 interface Props {
     legend: string;
@@ -7,12 +8,29 @@ interface Props {
     className?: string;
     helpLabel?: string;
     errorLabel?: string;
+    variant?: LabelVariant;
+    forceCompact?: boolean;
 }
 
-export const FieldGroup = ({ legend, className, children, helpLabel, errorLabel }: Props) => (
-    <fieldset className={`jkl-field-group ${className ? className : ""}`}>
-        <legend className="jkl-field-group__legend">{legend}</legend>
-        {children}
-        {(helpLabel || errorLabel) && <SupportLabel helpLabel={helpLabel} errorLabel={errorLabel} />}
-    </fieldset>
-);
+export const FieldGroup = ({
+    legend,
+    className,
+    children,
+    helpLabel,
+    errorLabel,
+    variant = "medium",
+    forceCompact,
+}: Props) => {
+    const componentClassName = "jkl-field-group".concat(className ? ` ${className}` : "");
+    return (
+        <fieldset className={componentClassName}>
+            <legend className="jkl-field-group__legend">
+                <Label variant={variant} forceCompact={forceCompact}>
+                    {legend}
+                </Label>
+            </legend>
+            {children}
+            {(helpLabel || errorLabel) && <SupportLabel helpLabel={helpLabel} errorLabel={errorLabel} />}
+        </fieldset>
+    );
+};