+++ packages/logo-react/src/LogoMain.tsx
@@ -2,10 +2,18 @@ import React from "react";
 
 interface Props {
     className?: string;
+    title?: string;
 }
 
-export const LogoMain = ({ className = "" }: Props) => (
-    <svg className={`jkl-logo ${className}`} viewBox="0 0 994 187" xmlns="http://www.w3.org/2000/svg">
+export const LogoMain = ({ className = "", title = "Fremtind" }: Props) => (
+    <svg
+        aria-labelledby="logoMainTitle"
+        role="img"
+        className={`jkl-logo ${className}`}
+        viewBox="0 0 994 187"
+        xmlns="http://www.w3.org/2000/svg"
+    >
+        <title id="logoMainTitle">{title}</title>
         <path
             stroke="currentColor"
             fill="currentColor"