+++ portal/src/components/Menu/Menu.tsx
@@ -1,23 +1,68 @@
-import React, { useState, useRef, useContext } from "react";
-import { Link } from "gatsby";
+import React, { useState, useRef, useContext, useEffect } from "react";
+import { Link, graphql, useStaticQuery } from "gatsby";
 import { LocationProvider } from "@reach/router";
 import { Accordion, AccordionItem } from "@fremtind/jkl-accordion-react";
 import { Hamburger } from "@fremtind/jkl-hamburger-react";
-import { coreLinks, developerLinks, designerLinks, componentLinks, profileLinks, exampleLinks } from "./links";
+import { ToggleSwitch } from "@fremtind/jkl-toggle-switch-react";
+import { ComponentDocPage, staticLinks, mainMenu } from "./links";
 import "@fremtind/jkl-accordion/accordion.min.css";
 import "@fremtind/jkl-hamburger/hamburger.min.css";
-import { ToggleSwitch } from "@fremtind/jkl-toggle-switch-react";
+import "@fremtind/jkl-text-input/text-input.min.css";
 import { ThemeContext } from "../Layout/Layout";
+import { FilterField } from "./FilterField";
 
 import "./Menu.scss";
 
 export function Menu() {
     const [showMenu, toggleShowMenu] = useState(false);
 
+    const { allSitePage: componentDocumentationPage, allMarkdownRemark } = useStaticQuery(graphql`
+        query getPages {
+            allSitePage(
+                filter: { path: { regex: "/-react/documentation/" } }
+                sort: { order: ASC, fields: context___frontmatter___title }
+            ) {
+                edges {
+                    node {
+                        id
+                        path
+                        context {
+                            frontmatter {
+                                title
+                                react
+                            }
+                        }
+                    }
+                }
+            }
+            allMarkdownRemark {
+                edges {
+                    node {
+                        frontmatter {
+                            path
+                            title
+                        }
+                    }
+                }
+            }
+        }
+    `);
     const menuRef = useRef(null);
 
     const toggleMenu = (show: boolean) => toggleShowMenu(show);
     const { theme, toggleTheme } = useContext(ThemeContext);
+
+    const [menu, setMenu] = useState(mainMenu(allMarkdownRemark, componentDocumentationPage));
+    const [filter, setFilter] = useState("");
+    useEffect(() => {
+        function filterComponents(filter: string) {
+            return componentDocumentationPage.edges.filter((edge: ComponentDocPage) => {
+                return edge.node.context.frontmatter.title.toLowerCase().includes(filter.toLowerCase());
+            });
+        }
+        const filteredComponents = filterComponents(filter);
+        setMenu(mainMenu(allMarkdownRemark, { edges: filteredComponents }));
+    }, [filter]);
 
     return (
         <LocationProvider>
@@ -35,121 +80,46 @@ export function Menu() {
 
                     <nav className="portal-menu__bar">
                         <Accordion>
-                            <AccordionItem
-                                title="Grunnleggende"
-                                startExpanded={!!location.pathname.match(/(core|jokul)/) || location.pathname === "/"}
-                            >
-                                {coreLinks.map((link) => (
-                                    <Link
-                                        key={link.title}
-                                        className="portal-menu__link"
-                                        to={`/${link.section}/${link.page}`}
-                                        tabIndex={showMenu ? 0 : -1}
-                                    >
-                                        {link.title}
-                                    </Link>
-                                ))}
-                            </AccordionItem>
-
-                            <AccordionItem
-                                title="Profilelementer"
-                                startExpanded={location.pathname.includes("profile")}
-                            >
-                                {profileLinks.map((link) => (
-                                    <Link
-                                        key={link.title}
-                                        className="portal-menu__link"
-                                        to={`/${link.section}/${link.page}`}
-                                        tabIndex={showMenu ? 0 : -1}
-                                    >
-                                        {link.title}
-                                    </Link>
-                                ))}
-                            </AccordionItem>
-                            <AccordionItem title="For designere" startExpanded={location.pathname.includes("designer")}>
-                                {designerLinks.map((link) => (
-                                    <Link
-                                        key={link.title}
-                                        className="portal-menu__link"
-                                        to={`/${link.section}/${link.page}`}
-                                        tabIndex={showMenu ? 0 : -1}
-                                    >
-                                        {link.title}
-                                    </Link>
-                                ))}
-                            </AccordionItem>
-                            <AccordionItem
-                                title="For utviklere"
-                                startExpanded={location.pathname.includes("developer")}
-                            >
-                                {developerLinks.map((link) => (
-                                    <Link
-                                        key={link.title}
-                                        className="portal-menu__link"
-                                        to={`/${link.section}/${link.page}`}
-                                        tabIndex={showMenu ? 0 : -1}
-                                    >
-                                        {link.title}
-                                    </Link>
-                                ))}
-                            </AccordionItem>
-                            <AccordionItem title="Komponenter" startExpanded={location.pathname.includes("components")}>
-                                {componentLinks.map((link) => (
-                                    <Link
-                                        key={link.title}
-                                        className="portal-menu__link"
-                                        to={`/${link.section}/${link.page}`}
-                                        tabIndex={showMenu ? 0 : -1}
-                                    >
-                                        {link.title}
-                                    </Link>
-                                ))}
-                            </AccordionItem>
-                            <AccordionItem title="Eksempel" startExpanded={location.pathname.includes("example")}>
-                                {exampleLinks.map((link) => (
-                                    <Link
-                                        key={link.title}
-                                        className="portal-menu__link"
-                                        to={`/${link.section}/${link.page}`}
-                                        tabIndex={showMenu ? 0 : -1}
-                                    >
-                                        {link.title}
-                                    </Link>
-                                ))}
-                            </AccordionItem>
+                            {menu.map(({ sectionTitle, pages, matchingLocation }) => (
+                                <AccordionItem
+                                    key={sectionTitle}
+                                    title={sectionTitle}
+                                    startExpanded={matchingLocation(location)}
+                                >
+                                    {sectionTitle === "Komponenter" && (
+                                        <FilterField filter={filter} setFilter={setFilter} />
+                                    )}
+                                    {pages.map(({ title, path }) => (
+                                        <Link
+                                            key={title}
+                                            className="portal-menu__link"
+                                            activeClassName="portal-menu__link--current"
+                                            to={path}
+                                            tabIndex={showMenu ? 0 : -1}
+                                        >
+                                            {title}
+                                        </Link>
+                                    ))}
+                                </AccordionItem>
+                            ))}
                         </Accordion>
-                        <a
-                            className="portal-menu__link portal-menu__link--github jkl-p"
-                            href="https://github.com/fremtind/jokul"
-                            tabIndex={showMenu ? 0 : -1}
-                        >
-                            Kode på Github
-                        </a>
-                        <a
-                            className="portal-menu__link portal-menu__link--github jkl-p"
-                            href="https://github.com/fremtind/jokul/issues/new?assignees=&labels=bug&template=bug_report.md&title="
-                            tabIndex={showMenu ? 0 : -1}
-                        >
-                            Rapporter feil
-                        </a>
-                        <a
-                            className="portal-menu__link portal-menu__link--github jkl-p"
-                            href="https://github.com/fremtind/jokul/issues/new?assignees=&labels=enhancement&template=feature_request.md&title="
-                            tabIndex={showMenu ? 0 : -1}
-                        >
-                            Forslag ny funksjon
-                        </a>
-                        <a
-                            className="portal-menu__link portal-menu__link--figma jkl-p"
-                            href="https://www.figma.com/file/TkbB9ANfejDSjB2u4u1OEuqM/J%C3%B8kul-components"
-                            tabIndex={showMenu ? 0 : -1}
-                        >
-                            Designbibliotek
-                        </a>
+
+                        {staticLinks.map(({ linkText, href, className }) => (
+                            <a
+                                className={`portal-menu__link jkl-p ${className}`}
+                                href={href}
+                                tabIndex={showMenu ? 0 : -1}
+                                key={linkText}
+                            >
+                                {linkText}
+                            </a>
+                        ))}
+
                         <ToggleSwitch
+                            inverted={theme === "dark"}
                             className="jkl-spacing--top-2 jkl-spacing--bottom-3"
                             checked={theme === "dark"}
-                            onChange={toggleTheme}
+                            onChange={(e) => toggleTheme(e.target.checked)}
                         >
                             Dark(beta)
                         </ToggleSwitch>