+++ portal/src/pages/index.tsx
@@ -1,10 +1,8 @@
 import React from "react";
 import { Layout } from "../components";
 import ReactMarkdown from "react-markdown";
-import { renderer } from "../presentation/markdownRenderer";
-// @ts-ignore
+import { jokulRenderer } from "../presentation/markdownRenderer";
 import frontPageMarkdown from "!raw-loader!../texts/FrontPage.md";
-// @ts-ignore
 import HeroMarkdown from "!raw-loader!../texts/frontpage/Hero.md";
 import "@fremtind/jkl-core/core.min.css";
 import "./style.scss";
@@ -13,10 +11,10 @@ const IndexPage = () => (
     <Layout showFooter={false}>
         <img className="portal-main__background" src="mountains.jpg" alt="Fjell illustasjon" />
         <div className="portal-main__hero">
-            <ReactMarkdown renderers={renderer} source={HeroMarkdown} />{" "}
+            <ReactMarkdown renderers={jokulRenderer} source={HeroMarkdown} />{" "}
         </div>
         <div className="portal-main__content">
-            <ReactMarkdown renderers={renderer} source={frontPageMarkdown} />
+            <ReactMarkdown renderers={jokulRenderer} source={frontPageMarkdown} />
         </div>
     </Layout>
 );