+++ portal/src/utils/copyToClipboard.ts
@@ -1,8 +1,0 @@
-export const copyToClipboard = (text: string) => {
-    const textField = document.createElement("textarea");
-    textField.innerText = text;
-    document.body.appendChild(textField);
-    textField.select();
-    document.execCommand("copy");
-    textField.remove();
-};