+++ packages/table-react/CHANGELOG.md
@@ -3,19 +3,83 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# 0.2.0 (2019-09-23)
+## 1.0.7 (2019-12-20)
+
+**Note:** Version bump only for package @fremtind/jkl-table-react
 
 
-### Bug Fixes
 
-* **jkl-table-react:** add default onRowClick with navigation ([01256b5](https://github.com/fremtind/jokul/commit/01256b5))
-* **jkl-table-react:** let the TableRow click handler handle navigation ([b94847a](https://github.com/fremtind/jokul/commit/b94847a))
-* **jkl-table-react:** use jkl-sr-only to hide anchors in tables ([6b79957](https://github.com/fremtind/jokul/commit/6b79957))
-* **jkl-table-react:** use th instead of td in table head row ([9228d0b](https://github.com/fremtind/jokul/commit/9228d0b))
 
+
+## 1.0.6 (2019-12-20)
+
+**Note:** Version bump only for package @fremtind/jkl-table-react
+
+
+
+
+
+## 1.0.5 (2019-12-20)
+
+**Note:** Version bump only for package @fremtind/jkl-table-react
+
+
+
+
+
+## 1.0.4 (2019-12-20)
+
+**Note:** Version bump only for package @fremtind/jkl-table-react
+
+
+
+
+
+## 1.0.3 (2019-12-20)
+
+**Note:** Version bump only for package @fremtind/jkl-table-react
+
+## 1.0.2 (2019-12-19)
+
+**Note:** Version bump only for package @fremtind/jkl-table-react
+
+## [1.0.1](https://github.com/fremtind/jokul/compare/@fremtind/jkl-table-react@1.0.0...@fremtind/jkl-table-react@1.0.1) (2019-12-17)
+
+**Note:** Version bump only for package @fremtind/jkl-table-react
+
+# [1.0.0](https://github.com/fremtind/jokul/compare/@fremtind/jkl-table-react@0.3.2...@fremtind/jkl-table-react@1.0.0) (2019-12-03)
 
 ### Features
 
-* **jkl-table:** change prop name for click a row in table ([5dfa8c9](https://github.com/fremtind/jokul/commit/5dfa8c9))
-* **jkl-table-react:** add Table react component ([8f3aaf0](https://github.com/fremtind/jokul/commit/8f3aaf0))
-* **jkl-table-react:** depend on correct jkl-core version ([229c99e](https://github.com/fremtind/jokul/commit/229c99e))
+-   **table-react:** change table row types to enum ([d9394a9](https://github.com/fremtind/jokul/commit/d9394a924c82f5d837a91a4cf008c9691b51567d))
+-   **table-react:** export table row types from index ([d4afa3b](https://github.com/fremtind/jokul/commit/d4afa3bac4add3090189e726346c67d59a89326c))
+-   **table-react:** use hrefLabel as title attribute on anchor rows ([e719622](https://github.com/fremtind/jokul/commit/e719622eeabf14fceb0b5382fc1e53b91d1bd1a4)), closes [#515](https://github.com/fremtind/jokul/issues/515)
+
+## [0.3.2](https://github.com/fremtind/jokul/compare/@fremtind/jkl-table-react@0.3.1...@fremtind/jkl-table-react@0.3.2) (2019-10-09)
+
+**Note:** Version bump only for package @fremtind/jkl-table-react
+
+## [0.3.1](https://github.com/fremtind/jokul/compare/@fremtind/jkl-table-react@0.3.0...@fremtind/jkl-table-react@0.3.1) (2019-09-26)
+
+**Note:** Version bump only for package @fremtind/jkl-table-react
+
+# [0.3.0](https://github.com/fremtind/jokul/compare/@fremtind/jkl-table-react@0.2.0...@fremtind/jkl-table-react@0.3.0) (2019-09-26)
+
+### Features
+
+-   **table-react:** add className to table ([d3d4e06](https://github.com/fremtind/jokul/commit/d3d4e06)), closes [#377](https://github.com/fremtind/jokul/issues/377)
+
+# 0.2.0 (2019-09-23)
+
+### Bug Fixes
+
+-   **jkl-table-react:** add default onRowClick with navigation ([01256b5](https://github.com/fremtind/jokul/commit/01256b5))
+-   **jkl-table-react:** let the TableRow click handler handle navigation ([b94847a](https://github.com/fremtind/jokul/commit/b94847a))
+-   **jkl-table-react:** use jkl-sr-only to hide anchors in tables ([6b79957](https://github.com/fremtind/jokul/commit/6b79957))
+-   **jkl-table-react:** use th instead of td in table head row ([9228d0b](https://github.com/fremtind/jokul/commit/9228d0b))
+
+### Features
+
+-   **jkl-table:** change prop name for click a row in table ([5dfa8c9](https://github.com/fremtind/jokul/commit/5dfa8c9))
+-   **jkl-table-react:** add Table react component ([8f3aaf0](https://github.com/fremtind/jokul/commit/8f3aaf0))
+-   **jkl-table-react:** depend on correct jkl-core version ([229c99e](https://github.com/fremtind/jokul/commit/229c99e))