+++ packages/table/CHANGELOG.md
@@ -3,19 +3,75 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# 0.2.0 (2019-09-23)
+## 1.0.6 (2019-12-20)
 
+**Note:** Version bump only for package @fremtind/jkl-table
+
+
+
+
+
+## 1.0.5 (2019-12-20)
+
+**Note:** Version bump only for package @fremtind/jkl-table
+
+
+
+
+
+## 1.0.4 (2019-12-20)
+
+**Note:** Version bump only for package @fremtind/jkl-table
+
+
+
+
+
+## 1.0.3 (2019-12-20)
+
+**Note:** Version bump only for package @fremtind/jkl-table
+
+
+
+
+
+## 1.0.2 (2019-12-20)
+
+**Note:** Version bump only for package @fremtind/jkl-table
+
+## 1.0.1 (2019-12-19)
+
+**Note:** Version bump only for package @fremtind/jkl-table
+
+# [1.0.0](https://github.com/fremtind/jokul/compare/@fremtind/jkl-table@0.3.0...@fremtind/jkl-table@1.0.0) (2019-12-03)
+
+**Note:** Version bump only for package @fremtind/jkl-table
+
+# [0.3.0](https://github.com/fremtind/jokul/compare/@fremtind/jkl-table@0.2.1...@fremtind/jkl-table@0.3.0) (2019-10-09)
 
 ### Bug Fixes
 
-* **jkl-table:** move hover/active state to anchor-rows only ([f04128e](https://github.com/fremtind/jokul/commit/f04128e))
-* **jkl-table-react:** use jkl-sr-only to hide anchors in tables ([6b79957](https://github.com/fremtind/jokul/commit/6b79957))
-
+-   **core:** add and use wcag-compliant error/focus colors ([c714b45](https://github.com/fremtind/jokul/commit/c714b45))
 
 ### Features
 
-* **jkl-table:** add initial version of table ([582b456](https://github.com/fremtind/jokul/commit/582b456))
-* **jkl-table:** add support for anchor/link-rows ([8294390](https://github.com/fremtind/jokul/commit/8294390))
-* **jkl-table:** make anchor cover entire table data cell ([4be0b9a](https://github.com/fremtind/jokul/commit/4be0b9a))
-* **jkl-table:** make design-adjustments ([2ad87d2](https://github.com/fremtind/jokul/commit/2ad87d2))
-* **jkl-table:** remove right alignment of last column ([052e26b](https://github.com/fremtind/jokul/commit/052e26b))
+-   **core:** add pure white as a supporting color ([1558fa9](https://github.com/fremtind/jokul/commit/1558fa9))
+
+## [0.2.1](https://github.com/fremtind/jokul/compare/@fremtind/jkl-table@0.2.0...@fremtind/jkl-table@0.2.1) (2019-09-26)
+
+**Note:** Version bump only for package @fremtind/jkl-table
+
+# 0.2.0 (2019-09-23)
+
+### Bug Fixes
+
+-   **jkl-table:** move hover/active state to anchor-rows only ([f04128e](https://github.com/fremtind/jokul/commit/f04128e))
+-   **jkl-table-react:** use jkl-sr-only to hide anchors in tables ([6b79957](https://github.com/fremtind/jokul/commit/6b79957))
+
+### Features
+
+-   **jkl-table:** add initial version of table ([582b456](https://github.com/fremtind/jokul/commit/582b456))
+-   **jkl-table:** add support for anchor/link-rows ([8294390](https://github.com/fremtind/jokul/commit/8294390))
+-   **jkl-table:** make anchor cover entire table data cell ([4be0b9a](https://github.com/fremtind/jokul/commit/4be0b9a))
+-   **jkl-table:** make design-adjustments ([2ad87d2](https://github.com/fremtind/jokul/commit/2ad87d2))
+-   **jkl-table:** remove right alignment of last column ([052e26b](https://github.com/fremtind/jokul/commit/052e26b))